Use smaller data types

+++ src/Device/HCSR04/HCSR04.hpp
@@ -56,7 +56,7 @@ namespace devuino::device
 		{
 			if (debounce)
 			{
-				const unsigned int iterations {10};
+				const uint8_t iterations {10};
 
 				Lenght total {};
